refactor: mqpublisher application renamed to msvc-attractions-alerting. Jira ticket ques6
