--- mq-publisher/src/main/java/com/quesocololand/mqpublisher/services/abstractions/RabbitPublisherService.java
@@ -1,8 +1,0 @@
-package com.quesocololand.mqpublisher.services.abstractions;
-
-import com.quesocololand.mqpublisher.models.VisitorCount;
-
-public interface RabbitPublisherService {
-    //Methods of RabbitPublisherService
-    void publish(VisitorCount message);
-}
