+++ msvc-attractions-alerting/.gitattributes
