+++ msvc-attractions-alerting/.gitignore
