+++ msvc-attractions-alerting/src/main/java/com/quesocololand/msvcattractionsalerting/MsvcAttractionsAlertingApplication.java
@@ -1,11 +1,11 @@
-package com.quesocololand.mqpublisher;
+package com.quesocololand.msvcattractionsalerting;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class MqPublisherApplication {
+public class MsvcAttractionsAlertingApplication {
 	public static void main(String[] args) {
-		SpringApplication.run(MqPublisherApplication.class, args);
+		SpringApplication.run(MsvcAttractionsAlertingApplication.class, args);
 	}
 }
