+++ msvc-attractions-alerting/src/main/java/com/quesocololand/msvcattractionsalerting/services/abstractions/RabbitVisitorCountsPublisherService.java
@@ -1,0 +1,8 @@
+package com.quesocololand.msvcattractionsalerting.services.abstractions;
+
+import com.quesocololand.msvcattractionsalerting.models.VisitorCount;
+
+public interface RabbitVisitorCountsPublisherService {
+    //Methods of RabbitVisitorCountsPublisherService
+    void publish(VisitorCount message);
+}
