+++ msvc-attractions-alerting/src/main/resources/application.properties
